Check limits and plan grants before automatic enrollments
